+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),  
 and this project ~adheres to [Semantic Versioning](https://semver.org/).
 
+## [1.0.2511.0413] - 2025-11-04
+
+### Fixed
+- Migration not honoring the selected host
+- Multiple netapp controllers causing confusion.
+
+
+
+## [1.0.2510.3113] - 2025-10-31
+
+### Fixed
+- Page Settings Netapp. Selected volumes not getting display as selected.
+
+
 
 ## [1.0.2510.3113] - 2025-10-31